Accordion toggle context created
Consumed by the header to conditionally render either the close or open button and also by the .Body component to return the Body or null.

Styled the imgs as well as added some horizontal padding to Inner and made the Header a flex container in order to space out the img and header text.