Fixed multiple features open simultaneously
The context provider now wraps around the Card.Group, providing each container with context now including the activeContainerIndex, set when clicking on a Card.Item.

Similar to the Accordion component, an index prop is passed to the Card.Item and Card.Feature components to be compared to the activeContainerIndex state var before rendering Feature.